backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
transformers
google-genai
python-dotenv
pytest
pytest-asyncio
httpx
